Change the DTMF pattern/regex
I did some tests and found I can get GPT to respond somewhat more reliably when I ask for the format `Pressing 2.` versus the original format of `*presses 2*`, especially when I set a lower temperature.  It's all probabilistic so it's hard to get it to work perfectly, but this approach seems a little better.